Change sound from int16 to float

+++ src/Client/EmulatorState.cpp
@@ -96,8 +96,8 @@ void SystemIO::blit(uint16_t x, uint16_t y, std::vector<uint8_t> buffer) {
     std::memcpy(ptr, buffer.data(), buffer.size());
 }
 
-void SystemIO::sound(int16_t frequency, uint16_t duration) {
-    soundBuffer.push(std::pair<int16_t, uint16_t>(frequency, duration));
+void SystemIO::sound(float frequency, uint16_t duration) {
+    soundBuffer.push(std::pair<float, uint16_t>(frequency, duration));
 }
 
 void EmulatorState::onRender(State *state, const uint32_t time) {
@@ -154,7 +154,7 @@ void EmulatorState::onTick(State *state, const uint32_t time) {
         }
     }
 
-    std::optional<std::pair<int16_t, uint16_t>> s = sysio->nextSound();
+    std::optional<std::pair<float, uint16_t>> s = sysio->nextSound();
     while (s != std::nullopt) {
         auto sound = *s;
         state->getSys()->sound(sound.first, sound.second);
