+++ src/Client/EmulatorState.h
@@ -22,7 +22,7 @@ namespace Client {
             std::array<uint8_t, 320*180> screen;
 
             std::queue<char> inputBuffer;
-            std::queue<std::pair<int16_t, uint16_t>> soundBuffer;
+            std::queue<std::pair<float, uint16_t>> soundBuffer;
         public:
             SystemIO();
 
@@ -58,13 +58,13 @@ namespace Client {
 
             void blit(uint16_t x, uint16_t y, std::vector<uint8_t> buffer);
 
-            void sound(int16_t frequency, uint16_t duration);
+            void sound(float frequency, uint16_t duration);
 
             std::array<uint8_t, 320*180> getScreen() {
                 return screen;
             }
 
-            std::optional<std::pair<int16_t, uint16_t>> nextSound() {
+            std::optional<std::pair<float, uint16_t>> nextSound() {
                 if (soundBuffer.size() == 0)
                     return std::nullopt;
