+++ src/Sys/SFML.h
@@ -44,7 +44,7 @@ namespace Sys {
                 window.setKeyRepeatEnabled(enable);
             }
 
-            void sound(int16_t frequency, uint16_t duration);
+            void sound(float frequency, uint16_t duration);
     };
 }; // Sys
